requirements.txt
langgraph
langchain-openai
langchain-core
qdrant-client
openai
python-dotenv
datasets
PyMuPDF
